requirements.txt
yfinance>=0.2.28
akshare>=1.17.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
matplotlib>=3.8.0
seaborn>=0.13.0